Start work on refactoring the way Event-listening threads work
The goal here is to move work out of listening threads and into the
main thread.  The hope is that this makes error handling easier,
fewer errors will occur in a subsidiary thread floating around
somewhere, and work happens in a somewhat more orderly, linear
fashion.  It also better reflects the reason we're even using
threads in the first place--to listen for I/O from multiple,
blocking sources.