Switch to an address-parsing function that resolves hostnames
Theoretically this ought to have been the end of issue #4, but,
it's not finished yet.  If a hostname parses successfully but
there isn't anyone listening on the first possible thing to try,
it can go through then fail with a Connection refused (os error
111) type issue.  The best way to fix this might be to start all
connections in the thread and just throw all of the parse results
into the reader thread.